refactor: header filters field
