+++ apps/blog/src/app/components/pages/editor/components/WordCountDisplay.tsx
@@ -1,0 +1,15 @@
+interface WordCountDisplayProps {
+  wordCount: number;
+  minutesToRead: number;
+}
+
+export function WordCountDisplay({
+  wordCount,
+  minutesToRead,
+}: WordCountDisplayProps) {
+  return (
+    <span className="text-muted-foreground text-sm">
+      {`${minutesToRead} min read`} ({wordCount} words)
+    </span>
+  );
+}
