+++ apps/blog/src/app/components/pages/editor/components/fields/TitleField.tsx
@@ -1,0 +1,34 @@
+import type { Control } from "react-hook-form";
+
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+  Input,
+} from "@ashgw/ui";
+
+import type { PostEditorDto } from "~/api/models/post";
+
+interface TitleFieldProps {
+  control: Control<PostEditorDto>;
+}
+
+export function TitleField({ control }: TitleFieldProps) {
+  return (
+    <FormField
+      control={control}
+      name="title"
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>Title</FormLabel>
+          <FormControl>
+            <Input placeholder="Blog Title" {...field} />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
